Adds swappable
