+++ tests/test_examples/test_pair1.py
@@ -2,7 +2,7 @@ from typing import Callable, Tuple, TypeVar
 
 from typing_extensions import final
 
-from returns.interfaces import bimappable, bindable, equable, lashable
+from returns.interfaces import bindable, equable, lashable, swappable
 from returns.primitives.container import BaseContainer, container_equality
 from returns.primitives.hkt import Kind2, SupportsKind2, dekind
 
@@ -18,7 +18,7 @@ class Pair(
     BaseContainer,
     SupportsKind2['Pair', _FirstType, _SecondType],
     bindable.Bindable2[_FirstType, _SecondType],
-    bimappable.BiMappable2[_FirstType, _SecondType],
+    swappable.Swappable2[_FirstType, _SecondType],
     lashable.Lashable2[_FirstType, _SecondType],
     equable.Equable,
 ):
@@ -79,7 +79,7 @@ class Pair(
     ) -> 'Pair[_FirstType, _NewSecondType]':
         return dekind(function(self._inner_value[1]))
 
-    # `BiMappableN` part:
+    # `SwappableN` part:
 
     def swap(self) -> 'Pair[_SecondType, _FirstType]':
         return Pair((self._inner_value[1], self._inner_value[0]))
