+++ tests/test_examples/test_pair4.py
@@ -4,7 +4,7 @@ from typing import Callable, ClassVar, NoReturn, Sequence, Tuple, Type, TypeVar
 from typing_extensions import final
 
 from returns.contrib.hypothesis.laws import check_all_laws
-from returns.interfaces import bimappable, bindable, equable, lashable
+from returns.interfaces import bindable, equable, lashable, swappable
 from returns.primitives.asserts import assert_equal
 from returns.primitives.container import BaseContainer, container_equality
 from returns.primitives.hkt import Kind2, KindN, SupportsKind2, dekind
@@ -50,7 +50,7 @@ class _LawSpec(LawSpecDef):
 
 class PairLikeN(
     bindable.BindableN[_FirstType, _SecondType, _ThirdType],
-    bimappable.BiMappableN[_FirstType, _SecondType, _ThirdType],
+    swappable.SwappableN[_FirstType, _SecondType, _ThirdType],
     lashable.LashableN[_FirstType, _SecondType, _ThirdType],
     equable.Equable,
 ):
@@ -186,7 +186,7 @@ class Pair(
         """
         return dekind(function(self._inner_value[1]))
 
-    # `BiMappableN` part:
+    # `SwappableN` part:
 
     def swap(self) -> 'Pair[_SecondType, _FirstType]':
         """
